move command from actions.js to queries.js
