+++ src/agent/commands/queries.js
@@ -2,6 +2,7 @@ import * as world from '../library/world.js';
 import * as mc from '../../utils/mcdata.js';
 import { getCommandDocs } from './index.js';
 import convoManager from '../conversation.js';
+import { load } from 'cheerio';
 
 const pad = (str) => {
     return '\n' + str + '\n';
@@ -213,6 +214,35 @@ export const queryList = [
             console.log(craftingPlan);
             return pad(craftingPlan);
         },
+    },
+    {
+        name: '!searchWiki',
+        description: 'Search the Minecraft Wiki for the given query.',
+        params: {
+            'query': { type: 'string', description: 'The query to search for.' }
+        },
+        perform: async function (agent, query) {
+            const url = `https://minecraft.wiki/w/${query}`
+            try {
+                const response = await fetch(url);
+                if (response.status === 404) {
+                  return `${query} was not found on the Minecraft Wiki. Try adjusting your search term.`;
+                }
+                const html = await response.text();
+                const $ = load(html);
+            
+                const parserOutput = $("div.mw-parser-output");
+                
+                parserOutput.find("table.navbox").remove();
+
+                const divContent = parserOutput.text();
+            
+                return divContent.trim();
+              } catch (error) {
+                console.error("Error fetching or parsing HTML:", error);
+                return `The following error occured: ${error}`
+              }
+        }
     },
     {
         name: '!help',
